% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkPC.R
\name{checkPC}
\alias{checkPC}
\title{Check Current PC}
\usage{
checkPC()
}
\value{
The name of the current computer
}
\description{
Function to check which computer I'm currently using, for adjusting directories.
}
\examples{
checkPC()

}

