% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chunk_it_up.R
\name{chunk_it_up}
\alias{chunk_it_up}
\title{Chunk It Up}
\usage{
chunk_it_up(.data, chunksize = NULL, nchunks = NULL)
}
\arguments{
\item{.data}{The data frame or tibble}

\item{chunksize}{The number of rows per chunk. Specify either chunksize or nchunks.}

\item{nchunks}{The number of chunks. Specify either nchunks or chunksize.}
}
\value{
A chunk number vector
}
\description{
Create a chunk column for parallelising tasks
}
\examples{
data(mtcars)

mtcars$chunk <- chunk_it_up(mtcars, nchunks=3)
mtcars$chunk <- chunk_it_up(mtcars, chunksize=5)
}
