% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clear.all.R
\name{clear.all}
\alias{clear.all}
\title{Clear all data from the workspace}
\usage{
clear.all()
}
\description{
Function to clear all data from the workspace, but leave functions intact.  It must be noted that it does not fully work, and one must run
the function without brackets and copy the function text to remove everything.
}
\examples{
clear.all

}

