% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initProjectFolder.R
\name{initProjectFolder}
\alias{initProjectFolder}
\title{Create a Reproducible Project Folder}
\usage{
initProjectFolder()
}
\value{
Nothing is returned. The folders and text files are created.
}
\description{
Function to create a set of working folders for a reproducible project
}
\examples{
initProjectFolder()

}
\author{
Granville J Matheson, \email{mathesong@gmail.com}
}
\references{
The folder structure comes from Karl Broman's webinar for the American Statistical
Association (ASA): https://www.amstat.org/asa/files/pdfs/EDU-ReproducibleResearchWebinarTranscript.pdf
}

