% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/limit_df.R
\name{limit_df}
\alias{limit_df}
\title{Limit a Data Frame}
\usage{
limit_df(df, limits, columns)
}
\arguments{
\item{df}{The data frame whose values are to be limited}

\item{limits}{The limits to be imposed on the column/columns of the data frame}
}
\value{
The data frame after being limited
}
\description{
Function to limit the values of a column or columns of a data frame
}
\examples{
a <- data.frame(Year = 1900:1980, Potatos = round(runif(81)*100))
limit_df(a, c(0, 50), 2)

}

