% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maxz.R
\name{maxz}
\alias{maxz}
\title{Maximum Normalised Z Value}
\usage{
maxz(values)
}
\arguments{
\item{values}{Vector of Values}
}
\description{
This function calculates the maximum normalised z values for a data set as well as the range.
}
\examples{
maxz( runif(20) )

}
