% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nearestN.R
\name{nearestN}
\alias{nearestN}
\title{Nearest Value Index}
\usage{
nearestN(value, values)
}
\arguments{
\item{value}{The value to be matched}

\item{values}{The set of values with which the value is being compared}
}
\value{
Index of the nearest value
}
\description{
Function to find the index of the nearest value from within a set of values
}
\examples{
a <- runif()
nearestN(50, a)

}
