% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nearestV.R
\name{nearestV}
\alias{nearestV}
\title{Nearest Value}
\usage{
nearestV(value, values)
}
\arguments{
\item{value}{The value to be matched}

\item{values}{The set of values with which the value is being compared}
}
\value{
The nearest value
}
\description{
Function to find the nearest value from within a set of values
}
\examples{
a <- runif()
nearestV(50, a)

}

