% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/percdata_lim.R
\name{percdata_lim}
\alias{percdata_lim}
\title{Percentage Data Limits}
\usage{
percdata_lim(
  x,
  y,
  perc,
  xlo.start = NULL,
  xhi.start = NULL,
  ylo.start = NULL,
  yhi.start = NULL
)
}
\arguments{
\item{x}{X values}

\item{y}{Y values}

\item{perc}{The percentage of points which should be included in the final figure}

\item{xlo.start}{Specify the starting parameter for xlo}

\item{xhi.start}{Specify the starting parameter for xhi}

\item{ylo.start}{Specify the starting parameter for ylo}

\item{yhi.start}{Specify the starting parameter for yhi}
}
\value{
Returns the x limits and the y limits for the new plot
}
\description{
Function to calculate axis limits for plots displaying large amounts of data with large spread.  
This function optimises the xlim and ylim parameters of the final plot in order to include at 
least a certain percentage of the data points, while minimising the extent of the x and y axes.
}
\examples{

x <- rnorm(1000)
y <- rnorm(1000)

percdata_lim(x,y,90)

}
