% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/residuals_for.R
\name{residuals_for}
\alias{residuals_for}
\title{Calculate residuals for a model excluding one of the predictors}
\usage{
residuals_for(model, predictor)
}
\arguments{
\item{model}{The input model, e.g. lm(a ~ b + c)}

\item{predictor}{The name of the predictor to be excluded}
}
\value{
A tibble containing true value, predicted values without the
  predictor, and the residuals without the predictor
}
\description{
Allows access to residuals without the influence of a particular predictor
}
\examples{

data(mtcars)
model1 <- lm(mpg ~ cyl + disp + gear, data=mtcars)
resout <- residuals_for(model1, "cyl")

model2 <- lm(residuals ~ cyl + 0, data=resout)

coef(model1)
coef(model2)


}
