% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scatter_heatmap.R
\name{scatter_heatmap}
\alias{scatter_heatmap}
\title{Scatter Heat Map}
\usage{
scatter_heatmap(xvalues, yvalues, xlabel, ylabel, xlimits, ylimits,
  alphaval = 1)
}
\arguments{
\item{xvalues}{X values}

\item{yvalues}{Y values}

\item{xlabel}{X axis label}

\item{ylabel}{Y axis label}

\item{xlimits}{Vector including limits of the x axis}

\item{ylimits}{Vector including limits of the x axis}

\item{alphaval}{The transparency of the points}
}
\value{
Returns the ggplot2 plot
}
\description{
Function to plot a scatter plot where each point is coloured by its 2-dimensional density.
Points beyond the limits will be sucked in to the edge values to depict that there exists more 
beyond the edge of the plot.
}
\examples{

x = rnorm(n = 50000, mean = 2, sd = 1)
y = rexp(50000, 1)

scatter_heatmap(x, y, 'X Values', 'Y Values', c(0, 3.5), c(-0.1, 8), 1 )

}

