% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sourceFolder.R
\name{sourceFolder}
\alias{sourceFolder}
\title{Source all R scripts within  folder}
\usage{
sourceFolder(folder)
}
\arguments{
\item{folder}{The folder to be sourced}
}
\description{
Function to source all the R scripts within a given folder
}
\examples{
sourceFolder('C:/Viruses')

}
