% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sumstats.R
\name{sumstats}
\alias{sumstats}
\title{Test Retest Summary Statistics
 
 This function calculates several common summary statistic parameters useful for test-retest studies.
 
 @param values Vector of values
 
 @examples
 sumstats( runif(20) )
 
 @export}
\usage{
sumstats(values)
}
\description{
Test Retest Summary Statistics
 
 This function calculates several common summary statistic parameters useful for test-retest studies.
 
 @param values Vector of values
 
 @examples
 sumstats( runif(20) )
 
 @export
}

