% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trt.R
\name{trt}
\alias{trt}
\title{Test Retest Summary}
\usage{
trt(values1, values2)
}
\arguments{
\item{values1}{Values from the first measurements}

\item{values2}{Values from the second measurements}
}
\description{
This function calculates several common parameters for evaluating test-retest reliability. The output
is a detailed list, containing one summarised tidy output called tidy.
}
\examples{
trt( runif(20), runif(20) )

}
