% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maxz.R
\name{maxz}
\alias{maxz}
\title{Test Retest Maximum Normalised Z Value
 
 This function calculates the maximum normalised z values from a test-retest data set. This is calculated
 
 @param values Vector of Values
 
 @examples
 maxz( runif(20) )
 @export}
\usage{
maxz(values)
}
\description{
Test Retest Maximum Normalised Z Value
 
 This function calculates the maximum normalised z values from a test-retest data set. This is calculated
 
 @param values Vector of Values
 
 @examples
 maxz( runif(20) )
 @export
}

