% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maxz.R
\name{maxz}
\alias{maxz}
\title{Maximum Normalised Z Value
 
 This function calculates the maximum normalised z values for a data set as well as the range.
 
 @param values Vector of Values
 
 @examples
 maxz( runif(20) )
 
 @export}
\usage{
maxz(values)
}
\description{
Maximum Normalised Z Value
 
 This function calculates the maximum normalised z values for a data set as well as the range.
 
 @param values Vector of Values
 
 @examples
 maxz( runif(20) )
 
 @export
}

