% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sumstats.R
\name{sumstats}
\alias{sumstats}
\title{Test Retest Summary Statistics}
\usage{
sumstats(values)
}
\arguments{
\item{values}{Vector of values}
}
\description{
This function calculates several common summary statistic parameters useful for test-retest studies.
}
\examples{
sumstats( runif(20) )

}

