% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trt.R
\name{trt}
\alias{trt}
\title{Test Retest Summary
 
 This function calculates several common parameters for evaluating test-retest reliability. The output
 is a detailed list, containing one summarised tidy output called tidy.
 
 @param values1 Values from the first measurements
 @param values2 Values from the second measurements
 
 @examples
 trt( runif(20), runif(20) )
 
 @export}
\usage{
trt(values1, values2)
}
\description{
Test Retest Summary
 
 This function calculates several common parameters for evaluating test-retest reliability. The output
 is a detailed list, containing one summarised tidy output called tidy.
 
 @param values1 Values from the first measurements
 @param values2 Values from the second measurements
 
 @examples
 trt( runif(20), runif(20) )
 
 @export
}

